+++ packages/dash-table/src/dash-table/derived/cell/wrappers.tsx
@@ -2,7 +2,7 @@ import * as R from 'ramda';
 import React from 'react';
 
 import { memoizeAll, memoizeOne } from 'core/memoizer';
-import { Data, IVisibleColumn, VisibleColumns, ColumnType, ActiveCell, SelectedCells, Datum, ColumnId, IViewportOffset } from 'dash-table/components/Table/props';
+import { Data, IVisibleColumn, VisibleColumns, ActiveCell, SelectedCells, Datum, ColumnId, IViewportOffset, Presentation } from 'dash-table/components/Table/props';
 import Cell from 'dash-table/components/Cell';
 import isActiveCell from 'dash-table/derived/cell/isActive';
 import isSelectedCell from 'dash-table/derived/cell/isSelected';
@@ -31,12 +31,14 @@ function getter(
                 const active = isActiveCell(activeCell, rowIndex + offset.rows, columnIndex + offset.columns);
                 const selected = isSelectedCell(selectedCells, rowIndex + offset.rows, columnIndex + offset.columns);
 
+                const isDropdown = column.presentation === Presentation.Dropdown;
+
                 const classes =
                     'dash-cell' +
                     ` column-${columnIndex}` +
                     (active ? ' focused' : '') +
                     (selected ? ' cell--selected' : '') +
-                    (column.type === ColumnType.Dropdown ? ' dropdown' : '');
+                    (isDropdown  ? ' dropdown' : '');
 
                 return elementCache([rowIndex, columnIndex], active, classes, columnIndex, column.id);
             },