+++ packages/dash-table/src/dash-table/handlers/cellEvents.ts
@@ -1,6 +1,7 @@
 import * as R from 'ramda';
 import { SelectedCells, ICellFactoryProps } from 'dash-table/components/Table/props';
 import isActive from 'dash-table/derived/cell/isActive';
+import reconcile from 'dash-table/reconcile';
 
 function isCellSelected(selectedCells: SelectedCells, idx: number, i: number) {
     return selectedCells && R.contains([idx, i], selectedCells);
@@ -98,9 +99,15 @@ export const handleChange = (propsFn: () => ICellFactoryProps, idx: number, i: n
         return;
     }
 
+    const result = reconcile(value, c);
+
+    if (!result.success) {
+        return;
+    }
+
     const newData = R.set(
         R.lensPath([realIdx, c.id]),
-        value,
+        result.value,
         data
     );
     setProps({