+++ packages/dash-table/src/dash-table/reconcile/number.ts
@@ -1,0 +1,17 @@
+import isNumeric from 'fast-isnumeric';
+
+import { INumberColumn } from 'dash-table/components/Table/props';
+import { reconcileNull, isNully } from './null';
+import { IReconciliation } from '.';
+
+export function coerce(value: any, options: INumberColumn | undefined): IReconciliation {
+    return isNumeric(value) ?
+        { success: true, value: +value } :
+        reconcileNull(value, options);
+}
+
+export function validate(value: any, options: INumberColumn | undefined): IReconciliation {
+    return typeof value === 'number' && !isNully(value) ?
+        { success: true, value } :
+        reconcileNull(value, options);
+}