+++ packages/dash-table/tests/visual/percy-storybook/fixtures.ts
@@ -1,3 +1,5 @@
+import { ColumnType, Presentation } from 'dash-table/components/Table/props';
+
 export default [
     {
         name: 'fixed rows -> dropdown',
@@ -7,7 +9,8 @@ export default [
                 {
                     name: 'Column 1',
                     id: 'column-1',
-                    type: 'dropdown',
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown,
                     options: [
                         {
                             label: 'Montréal',
@@ -132,7 +135,8 @@ export default [
                 {
                     name: 'Column 1',
                     id: 'column-1',
-                    type: 'dropdown',
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown,
                     options: [
                         {
                             label: 'Montréal',
@@ -168,7 +172,8 @@ export default [
                 {
                     name: 'Column 1',
                     id: 'column-1',
-                    type: 'dropdown',
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown,
                     options: [
                         {
                             label: 'Montréal',
@@ -183,7 +188,8 @@ export default [
                 {
                     name: 'Column 2',
                     id: 'column-2',
-                    type: 'dropdown',
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown,
                     options: [
                         {
                             label: 'Montréal',
@@ -198,7 +204,8 @@ export default [
                 {
                     name: 'Column 3',
                     id: 'column-3',
-                    type: 'dropdown',
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown,
                     options: [
                         {
                             label: 'Montréal',
@@ -248,7 +255,8 @@ export default [
                 {
                     name: ['Region', ''],
                     id: 'region',
-                    type: 'dropdown',
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown,
                     options: [
                         {
                             label: 'Hawaii',
@@ -490,7 +498,8 @@ export default [
                 {
                     id: 'column-1',
                     name: 'Column 1',
-                    type: 'dropdown'
+                    type: ColumnType.Text,
+                    presentation: Presentation.Dropdown
                 },
                 {
                     id: 'column-2',